+++ docs/scripts/constants/vendors.constant.js
@@ -1,0 +1,17 @@
+/**
+ * Generated header by C0ZEN for 21-days project
+ * Generated file vendors.constant on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 30/12/2017
+ * Time: 11:17
+ * Version: 1.0.0
+ */
+(function (angular) {
+	'use strict';
+
+	angular
+		.module('21days')
+		.constant('moment', moment);
+
+}(window.angular));